+++ src/runtime/dumps/index.ts
@@ -4,7 +4,7 @@
 
 export {
   buildRuntimeDumpsUserQuery,
-  getRuntimeDumpByUri,
+  getRuntimeDumpById,
   type IRuntimeDumpsListOptions,
   listRuntimeDumps,
   listRuntimeDumpsByUser,